🚚 reorganize atom.io/src/internal
