+++ packages/atom.io/src/internal/selector/index.ts
@@ -1,4 +1,5 @@
 export * from "./lookup-selector-sources"
 export * from "./register-selector"
+export * from "./selector-internal"
 export * from "./trace-selector-atoms"
 export * from "./update-selector-atoms"
