+++ packages/atom.io/src/internal/subscribe/index.ts
@@ -1,0 +1,1 @@
+export * from "./subscribe-to-root-atoms"
