+++ packages/atom.io/src/internal/subscribe/recall-state.ts
@@ -1,0 +1,19 @@
+import type { Atom } from "../atom"
+import type { ReadonlySelector, Selector } from "../selector"
+import type { Store } from "../store"
+import { IMPLICIT } from "../store"
+import { target } from "../transaction"
+
+export const recallState = <T>(
+	state: Atom<T> | ReadonlySelector<T> | Selector<T>,
+	store: Store = IMPLICIT.STORE,
+): T => {
+	const core = target(store)
+	if (!core.operation.open) {
+		store.config.logger?.warn(
+			`recall called outside of an operation. This is probably a bug.`,
+		)
+		return core.valueMap.get(state.key)
+	}
+	return core.operation.prev.get(state.key)
+}
