+++ packages/atom.io/src/internal/timeline/index.ts
@@ -1,1 +1,3 @@
 export * from "./add-atom-to-timeline"
+export * from "./time-travel-internal"
+export * from "./timeline-internal"
